pyproject.toml
[tool.poetry]
name = "pydwt"
version = "0.1.1a0"
description = ""
authors = ["gonza <matgonzalez@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
networkx = "^3.0"
matplotlib = "^3.6.3"
sqlalchemy = "^2.0.1"
dill = "^0.3.6"
typer = "^0.7.0"
pyyaml = "^6.0"
dependency-injector = "^4.41.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^23.1.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pydwt = "pydwt.app:app"